Get access_token from git_auth_token

+++ core/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, redirect
 from django.conf import settings
 import requests
-from rest_framework_simplejwt.tokens import RefreshToken
+from rest_framework_simplejwt.tokens import RefreshToken, AccessToken
 from django.contrib.auth.models import User
 from django.http import JsonResponse
 from rest_framework.response import Response
@@ -38,8 +38,10 @@ def git_auth_token(request):
     
     token_data = token_response.json()
     access_token = token_data.get('access_token')  # Pega o access token da resposta
+
     if not access_token:
         return JsonResponse({'error': 'Failed to obtain access token from GitHub'}, status=400)
+    request.session['jwt_token'] = access_token
 
     # Chama a função que cria o usuário com base no access token
     return create_user(request, access_token)
@@ -78,6 +80,7 @@ def create_user(request, access_token):
     refresh = RefreshToken.for_user(user)
     access_jwt = str(refresh.access_token)
     refresh_jwt = str(refresh)
+    
 
     # Retorna os dados do usuário e tokens (se quiser, pode incluir os tokens também)
     return JsonResponse({
@@ -85,20 +88,24 @@ def create_user(request, access_token):
         'email': email,
 
     })
-    login(request,user)
 
+def delete_user(request):
+    access_token = request.session.get('jwt_token') 
 
-def delete_user(request,access_token):
     user_response = requests.get(
         "https://api.github.com/user",
         headers={'Authorization': f'token {access_token}'}
     )
+
+    if user_response.status_code != 200:
+        return JsonResponse({'error': 'Failed to fetch user data from GitHub'}, status=400)
+
     user_data = user_response.json()
     username = user_data.get('login')
 
     try:
         user = User.objects.get(username=username)
         user.delete()
-        return JsonResponse({"Usuário deletado com sucesso.": 1})
+        return JsonResponse({'message': f'User {username} deleted successfully.'})
     except User.DoesNotExist:
-        return JsonResponse({"Usuário nao deletado com sucesso.":2})
+        return JsonResponse({'error': f'User {username} not found.'}, status=404)
